requirements.txt
fastapi==0.133.0
uvicorn==0.41.0
pydantic==2.12.5
sqlalchemy==2.0.46
numpy==1.26.4
scipy==1.12.0
pymoo==0.6.1
PyYAML==6.0.3
streamlit==1.54.0
plotly==5.24.1
requests==2.32.5
pytest==8.4.2
hypothesis==6.151.9
pytest-cov==7.0.0
httpx==0.28.1
watchdog==3.0.0
